PPO params change to args

+++ test_ppo_parallel.py
@@ -4,7 +4,7 @@ from ppo import ppo_run
 
 
 lambdas = [0.95]  # example with more than one lambda
-seeds = np.arange(0,5,1)
+seeds = np.arange(0,5,1, dtype=int)
 env = "CartPole-v1"
 
 # Create all combinations of lambda and seed
